requirements2.txt
asgiref==3.7.2
attrs==21.4.0
Automat==20.2.0
bcrypt==4.1.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==5.0.2
django-cors-headers==4.2.0
django-import-export==2.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.6.0
future==0.18.2
gunicorn==21.2.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
import-export==0.2.67.dev6
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.1
joblib==1.1.0
lxml==4.9.0
MarkupPy==1.14
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.2
pandas==1.4.2
paramiko==3.4.0
parsel==1.6.0
pefile==2021.9.3
Pillow==9.1.1
platformdirs==2.5.2
Protego==0.2.1
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pygame==2.1.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
scikit-learn==1.1.1
scipy==1.8.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
sshtunnel==0.4.0
tablib==3.2.1
threadpoolctl==3.1.0
tldextract==3.3.0
tqdm==4.64.0
Twisted==22.4.0
typing_extensions==4.9.0
tzdata==2022.1
ultimate-sitemap-parser==0.5
urllib3==1.26.9
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
w3lib==1.22.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
